修正了玩家数值无效的bug，现在修改表可以直接调了
Former-commit-id: 52cf73b65fcbd499fce42cc439a4223d97328b30

+++ Assets/Scripts/MagicianBase.cs
@@ -11,7 +11,7 @@ public class MagicianBase
     public float moveSpeed;         //移动速度
     public float damge;             //伤害
     public float bulletSpeed;       //子弹速度
-    public int fireSpeed;           //射速
+    public float fireSpeed;           //射速
     public List<bulletType> bulletEffect;  //子弹附加效果
     public GameObject bulletObj;    //子弹预制体
